Improve examples and remove print_with_headers() from API
The function `print_with_headers()` always felt a bit awkward and
inelegant. Move it to the bin instead. cargo public-items will start to
format its output as Markdown and will stop using it.

For quick and dirty output, it is sufficient to pretty-print the
`PublicItemsDiff` struct with "{:#?}".
